requirements.txt
anaconda==0.0.1.1
cryptography==41.0.1
fastapi==0.98.0
pandas==2.0.2
passlib==1.7.4
pydantic==1.10.9
PyJWT==2.7.0
requests==2.28.1
scikit-learn==1.2.2
scipy==1.11.0
starlette==0.27.0
streamlit==1.24.0
uvicorn==0.22.0

